Rename unittest folder
